Add demo setting to stress test movableContentOf
Adds a new MovableContentOfEverything setting to the
integration-tests/demos project to stress test movableContentOf
The goal here is to provide a quick way to verify how
functionality works when moving around inside a movableContentOf.

The default behavior is "Disabled", which will leave the content in a
single place (which is close to a no-op).

"Sibling" will swap the entire content of the app every 3 seconds,
between sibling positions:

```
if (flag) {
    content()
} else {
    content()
}
```

"BetweenSubcomposition" will swap the entire content of the app every 3
seconds, in and out of subcomposition:

```
if (flag) {
    content()
} else {
    BoxWithConstraints {
        content()
    }
}
```
With initial testing, there appears to be some issues with focus and
overscroll with the non-Disabled modes.

Test: N/A
Change-Id: If35e0e9b2d5b7e6cd709a32eda37d5436291d7c8