Add compat API to set disjoint parent of overlays
This CL adds an API for the first automatically-set disjoint parent in
AndroidX. When transition needs to add a view overlay, it will call into
this new method to set the disjoint parent, which fixes an issue that
prevents the correct resolution of view tree owners.

Moving forward, all overlays added to the hierarchy should use this new
compat method.

Test: ViewCompatTest, ViewTreeTest
Relnote: """
    Added `ViewCompat.addViewOverlay(View)` as an alternative to
    `View.getOverlay().add(View)`. The ViewCompat variant fixes an issue
    where owners like the LifecycleOwner and ViewModelStoreOwner could
    not be resolved between the view being overlaid and the overlay
    owner.
    """
Change-Id: I81413883e41e56e5ea76372debbd5b012c345f20