Introduce APIs to ensure the system bar contrast
The ContrastProtection is to ensure the readability of the elements on
the system bars. It describes a relative clean layer used to be placed
underneath the system bar. The derived class would describe how the
protection should be drawn.

The SystemBarStateMonitor provides necessary information of system bars
that need to be protected by ContrastProtections.

The ProtectionView is the concrete layer which draws a given list of
ContrastProtections. The caller can attach it to an area which is
complex on the edge of the window.

This CL also:
- adds WindowInsetsCompat.Side APIs for the caller to specify the sides.
- refines WindowInsetsAnimationCompat that:
  - it can return correct alpha, and
  - the interpolator and the duration of the legacy system bar
    animations follow the legacy specs

Relnote: Add a set of APIs about protecting system bars
Bug: 349980530
Test: Test with ProtectionComposableActivity in different combinations
      of navigation modes and orientations.
Change-Id: I9849c8cecab425d353375064e6a7e07cc8bb4b82