Add UserStyle support for ColorListUserStyleSetting
Relnote: Watch faces commonly allow the user to select colors using a ListUserStyle, with an icon for each ListOption. Since UserStyle schemas are sent over bluetooth, it is important to keep the size of the schema down which can be a problem if many dozens of color options are given due to all those icons. To help with this we've added ColorUserStyleSetting where the option contains a list of colors instead of an icon, which is much more compact.
Test: Manual testing
Bug: 367683475
Change-Id: Ib542e56a78ed0729795319a5b48447407cafc3a7