Add screenshot tests for Material 3
Adds initial setup and a few test cases using available components
(PrimaryLayout, ButtonGroup, scope with defaults, Text).

Not adding all variations of all colors and Text typographies as many of
them will be covered bby the new components and we shouldn't have a huge
amount of screenshot goldens here as well.

This also changes "size" from 390px to 525 as 390 corresponds to 144dp
which is too small.

Fixes: 372069575
Fixes: 358635507
Test: Added
Change-Id: I10a009eb633076608d1d046b2cc3177b4ab9c168