Update colors tokens API to be aligned with Compose and ColorScheme
ColorScheme object now holds the actual values for colors, and not the
enum map per tokens. This allows for much easier customization and
reading of the colors both in and outside of Material.

Shape will be update in a follow up.

Typography will stay as enums, as there's not 1 common object for all
parameters, and Typography is not customizable at all.

Bug: 369350414
Fixes: 372301651
Test: Updated
Relnote: "Updated Material3 colors to include ColorScheme concept, same as in Wear Compose."
Change-Id: If645e30349a8fb5372678903069da00f744bdc9e