Replace controlHints with controllerParameters
We might provide more informations when creating dynamic group route controller. The controlHints could be one of the controller parameters. But we might add more parameters in the future. To avoid repeatly updating the API surface, we create a ControllerParameters class to hold all parameters when creating route controllers.

Relnote: "New MediaRouteProvider APIs to create route controllers with controllerParameters"

Bug: 353645968
Test: Manually tested and unit tests
Change-Id: I703b9011c8ab76fd9849362f8805340dc4cc4c11