Introduce SessionConstants to SandboxedUiAdapter.Session
Introduces the concept of a session constants class to sessions created
by SandboxedUiAdapter. These constants represent data about the session
that will not change throughout the lifetime of the session. This change
does not introduce any new data to the openSession call, as only the
existing windowInputToken value has been added to SessionConstants in
this initial change.

Test: manual, SandboxedSdkViewTest, IntegrationTests
Bug: 373805386
Relnote: Added session constants to SandboxedUiAdapter sessions
Change-Id: I2843561492332167c544cdbaa25748fb1cfb0233