Change the snapshot id to be snapshotId: SnapshotId
The snapshot `id` can overflow in some extreme situations. The type
changed to be a `SnapshotId`, which is an expect/actual type that
is `Long` in the JVM. Using expect/actual allows the type  to be a
more appropriate type for other platforms that don't have a native
`Long` type. Using a `Long` on the JVM will will ensure it doesn't
overflow in any reasonable timeframe.

Fixes: 374821679
Test: existing tests, ./gradlew :compose:r:r:test
Relnote: """Added `snapshotId`, of type `Long`, to `Snapshot` and
derecated `id`.

The ID of a snapshot changed from a `Int` to a `Long` to avoid the
snapshot ID from overflowing on systems with very high frame rates
and long running animations.

The type was made expect/actual to allow this type to be `Int`,
or other type, on platforms that don't have a native `Long` type.

Platforms that do not have a native `Long`, such as JavaScript,
should avoid high frame rates (over 1000 FPS) which would cause
an overflow for `Int` approximate every 24 days.
"""

Change-Id: I38ac3d43d3af7f691c957017eaf6895f5ae6ebb9