Add a drop-down menu for ad formats to the manual test app
A new drop-down list is added to the options menu to choose between
banner and native ads. Native ad support will be added in the following
CLs, hence this option is disabled at the moment.

To simplify enabling/disabling of items in drop-down lists, a new
ArrayAdapter, DisabledItemsArrayAdapter, is added to the app. It allows
to explicitly set which items should be disabled in a list.

Bug: 369355774
Test: build, manual

Change-Id: I0a6a0e68628de76b97043b827764d466c11c29d9