Adding API for floating toolbar with adjacent FAB
This CL adds an API for a floating toolbar with an adjacent FAB.
The API supports horizontal and vertical layouts, as well as optional
positioning of the FAB to appear at the start or at the end of the
toolbar.

Also:
- Created custom layouts for the toolbar's variations and apply motion
  to show or hide the toolbar when the `expanded` state is changed.
- Adjusted the existing floating appbar tests to be parameterized
- Adjusted and added benchmark tests
- Added floating toolbar tokens file (needs updating)

Bug: 373460660
Test: Added tests and benchmark tests
Relnote: "Introducing a new API for creating a floating toolbar with an
attached Floating Action Button (FAB). The API provides flexible
customization options, allowing you to arrange the toolbar horizontally
or vertically and place the FAB at the start or end of the toolbar."

Change-Id: I9e3508dd059b8f6d4c5fd220d9b756ea4829d962