Reset semanticsConfiguration for deactivated nodes
Users can manually call invalidateSemantics() when they implement SemanticsModifierNode.
This can leave a node in a state with isSemanticsInvalidated == true even when the
node is deactivated.

This was not an issue before aosp/3296711 because we had a check in the semantics configuration
getter that returned null whenever a node was deactivated. We removed this check because it could
have unintended consequences and hide bugs.

This CL resets the isSemanticsInvalidated flag, and resets semantics configuration to null whenever
a layout node is deactivated.

Bug: 374823219
Bug: 375219962
Test: Added new tests
Change-Id: Ic634ebd66e464d2f7b30db48ec11d052f69fb6d5