Fix: Prevent App Crash on Clicking FAB After File Deletion
When a file is deleted from another app while it is open in the PDF viewer, the app crashes when FAB is clicked.
This CL ensures that if a file is deleted, it will no longer be displayed upon returning to the viewer. Instead, an error dialog will inform the user that the PDF file cannot be opened.

Changes:
1. Early return from the onStart() method if the file no longer exists.
2. Display an error view to notify the user.

Bug: 373297914
Test: build and run on test device
Change-Id: Ic6585c44bcb73649210496c716ea4c7107233224