Add new pdf-document-service artifact
This artifact will introduce a bound service to sandbox the rendering of PDF documents using the AOSP `PdfRenderer`, which utilizes the PDFium library. The service will only be used by the app itself and is not exposed to external callers. The `PdfRenderer` is not bundled within the library. This service will be sandboxed to ensure it only receives data passed from the app process, enhancing security and isolating potential vulnerabilities within PDFium. The implementation of this abstracted layer will be part of this artifact.

Relnote: Introduce new artifact to host Sandboxed PdfDocumentService
Test: ./gradlew :pdf:pdf-document-service:assemble
Bug: 377444646

Change-Id: I921fcbd54fa805d9f6e4e5084c54cf857fe1aa9b