Change Wear foundation pager snapAnimationSpec
Currently the snapAnimationSpec is set to a tween with a duration
of 150ms. This was set to ensure the Pager settles quickly so that
contents are focused and clickable, see b/345191109.

However a tween animation causes an ugly double jump when flinging
fast to the top or bottom of a SLC inside of a VerticalPager, see
b/349781047.

Changing to a Spring animation with no bounciness and high
stiffness achieves the same behaviour of settling quickly, and
avoids the double jump. Added a test to ensure this is the case.

Test to ensure the Pager settles quickly can be ran locally.

Bug: 303807950
Bug: 345191109
Bug: 350466286
Bug: 349781047
Test: Added test case to PagerTest.kt
Relnote: "Change pager snapAnimationSpec from Tween to Spring"

Change-Id: I10d0275b7f3b957af279893d2ed52d63e42b8115