Add `DelegatableNode.registerOnGlobalLayoutListener` that allows calculating occlusions
Adds an API for DelegatableNode that allows calculating occlusions from RectManager, using ThrottdledCallbacks.

Allows a user to get effectively the same Coordinate information as onRectChanged, but it also gives the user the chance to compute occlusions for the current state of the Layout at the moment of execution, returning a list of overlapping IntRect.

Relnote: "You may now calculate Composable occlusions with `RectInfo.calculateOcclusions()`."
Relnote: "Added extension function on `DelegatableNode` to register a listener for global layout changes."

Bug: 376110778
Test: OnGlobalLayoutListenerTest
Change-Id: I68b59e47e171889766c809343a0126844c66d905