Stylus hover icon modifier
This modifier is similar to the existing pointerHoverIcon modifier but
applies to stylus hover pointer input.

This also fixes a bug where a stylus hovering over an element with a
pointerHoverIcon modifier would cause the mouse pointer icon to change,
even though the mouse pointer icon is not shown for stylus hover.

Bug: 331289114
Bug: 337064733
Test: StylusHoverIconTest
Relnote: Adds stylusHoverIcon modifier.
Change-Id: Iff20af03b451228bc44c4356d4b4e89956cc3b45