Fix presenter race between concensecutive submitData
When AsyncPagingDataDiffer calculates diff, it temporarily sets the previousPrensenter to the previous list to ensure correct diffing. It resets the previousPresenter to null right after the coroutine for diffing completes. However, in the event of consecutive submitData where the first one is cancelled, the first presenter may return right after diff completes and never resets previousPresenter to null.

This leads to the bug where all calls to the presenter i.e. itemCount, getItem etc. are delegated to the previous list instead of the current list.

The previousPresenter is now reset within the diffing coroutine to ensure its proper state even if the presentation gets interrupted.

Test: ./gradlew paging:paging-runtime:cC
Bug: 371838050
Change-Id: I569f2872296aed1d36ae712bfd681ce6f65e677d