Fix IndexOutOfBound crash with link annotation
When we create nodes for link annotations, we should only be looking within the last visible line and not the full text.

In findParagraphByIndex we need to make sure that the index is within the visible lines. For that we check that it's not greater than last line's end.

In this change compared to the previous attempt to fix this issue (aosp/3320137) I remove an early return to keep the link's Box composable in composition since early return causes all sorts of issues, e.g. the retaining of the focus.

Bug: 372390054
Fixes: 374115892
Test: new tests + demo
Change-Id: Ic96d2868e3edd1ca11d718670fd494a8d402e086