Add `getMutableStateFlow` to `SavedStateHandle`
`SavedStateHandle.getMutableStateFlow(key): MutableStateFlow` function allows developers to obtain a `MutableStateFlow` associated with a given `key`, providing a convenient way to observe and modify state changes within the `SavedStateHandle` using `MutableStateFlow.update {}`.

A key advantage of `MutableStateFlow` is its availability across all KMP platforms, unlike `MutableLiveData`, which is limited to Android.

Please note that `getMutableStateFlow` is mutually exclusive with `getLiveData` for the same key. Using both methods with the same key will result in an `IllegalArgumentException`. This is because synchronizing the value between a `MutableLiveData` and a `MutableStateFlow` is not possible.

Please note that I have included a unit test for `savedStateProvider` as there appear to be none.

RelNote: "Add 'getMutableStateFlow' to 'SavedStateHandle'."
Test: SavedStateTest
Bug: 375408415
Change-Id: I04a4f0b470530a8993b467858abc45ecd2b42527