Add optimization for some collection types
Here we use SavedState's API for collections to optimize the following 9 types:

* List<Int>
* List<String>
* IntArray
* LongArray
* FloatArray
* DoubleArray
* CharArray
* BooleanArray
* Array<String>

Now these types are stored directly in the output SavedState and don't need to be decomposed or composed by our encoder and decoder. This can largely reduce the payload size of these types because we can get rid of the String keys and type info for each element in the collection. For example, when storing an Int array of 100 elements the number of bytes reduced from 1972 to 428 in the output Parcel on Android.

Note that when the collection is at the root the represenation no longer removes the nesting. For example, when encoding `intArrayOf(3, 5, 7)`, the representation in the output SavedState becomes:

```
{ "" = [3, 5, 7] }
```

instead of:

```
{ "0" = 3, "1" = 5, "2" = 7 }
```

Bug: 376025635
Test: SavedStateCodecTest.kt
Change-Id: I2cd1a29836382ae3208d650010562112ee90476b