Invoke View#onApplyWindowInsets via View#dispatchApplyWindowInsets
In this way, PFLAG3_APPLYING_INSETS will be applied, and the view won't
dispatch insets on its own. Otherwise, it can be redundant.

This CL also makes ViewCompat#dispatchApplyWindowInsets calls
ViewGroupCompat#dispatchApplyWindowInsets if comaptInsetsDispatch is
installed. Because an app might install comaptInsetsDispatch at
decorView, but some library components (such as ActionBarOverlayLayout)
might just dispatch insets to a descendant of it. In such cases, the
compatible dispatching should be still expected.

Fix: 376142568
Test: ViewGroupCompatTest
Change-Id: I7aec5a1d007a094088584f05db660cb39255c3c9