Refactor Autofill manager from expect/actual to interface
This CL also lifts the `requestAutofillForActiveElement` into common. Since the text toolbar changes in aosp/3312341 were added when `requestAutofill` was an Android-only API, this CL also removes all the logic added in that CL that was needed when the API was not in common.

See comments in aosp/3313535 for more notes.

Test: ensure existing tests pass
Fixes: 376080755
RelNote: "Changes Autofill manager to be an interface."
Change-Id: I8491407de5699bf7273b9f88bda11cc438e2fd62