Implement hashCode and equals for operator nodes.
To make retrieving the index of a node from a list of nodes easier,
we want to implement hashCode and equals so that users can get the index
by constructing the Node they are searching by and pass it into
getFirstIndexOf.

Bug: 356661489
Bug: 356662352
Bug: 369826552

Relnote: Implemented equals and hashCode for operator nodes.
Test: "Add test to check equals and hashCode return true for identical
nodes."

Change-Id: I1c7da1e709588be4e606c4175b85e2ad0d11e537