[Tooltip] Add onDismissRequest for tooltips.
Add onDismissRequest slot for tooltips, so makers can decide what happens after a user clicks outside of the tooltip.

Test: N/A
Bug: 375191227
Relnote: added a tooltip API that has a onDismissRequest parameter so makers can decide what happens when a user clicks outside of the tooltip. Deprecated the old API that doesn't contain this new parameter.
Change-Id: I99aef78418a6bb25f499bd67fac956a2c87be72e