Disable Paging logs on android unit tests
Non-kmp projects pulling in Paging 3.3 or later will automatically pull in android dependencies for android unit tests. This is by nature of how multiplatform sources are resolved.

Paging logger is an android dependency that requires instrumented tests or robolectric. As such it throws in non-instrumented unit tests. This CL checks os BUILD id and disables logging if null. This does not affect instrumented tests or robolectric tests as the ID in either case is non-null.

Test: existing tests pass
Bug: 331684448
Relnote: "Android unit tests pulling in Paging 3.3 or later will no longer throw from PagingLogger"
Change-Id: Ia9400db995e7f0357cae2647cbd2cef9cf6c3398