Removing recursion from currentSemanticsNodes setter
From the bug report, it looks like the semantics tree is invalidated
while the traversalOrder is being calculated.

This CL moves the code that calculates traversal order out of the class, and now as a static function that uses a reference to the semantics tree. The code calculating traversalOrder will not be affected by any invalidation to the semantics tree, as an invalidation will cause a new tree to be created which will have a separate path to calculate the new traversalOrder.

Test: No test as we don't have a repro, this is a speculative fix.
Bug: 375509809
Change-Id: Ide1d7f154d6c5097a6aa2ef85daeef62e34549f7