Update project creator instructions to use script wrapper
We've had a few bugs filed recently due to difficulty running the project creator in the correct environment, so I added a script wrapper to handle virtualenv creation / clean up and installation of necessary tools (toml, distutils).

The script will prompt you to install virtualenv if it is not already installed, but I left that part in the instructions because it's better to know earlier.

PiperOrigin-RevId: 693078661
Change-Id: I5d7b0288a8db461241d8eb2ca6aa20f597fb166c