Fix memory leak in Owner.layoutNodes map object
When a LayoutNode is re-used (in a LazyList) its semantics Id is updated, leaving the Owner mapping stale.

Meaning that whenever a LayoutNode that was re-used is detached, the actual LayoutNode object would remain in the map, with an Id that is no longer reachable (as they are generated incrementally).

This change makes sure the mapping object stays up to date with the change. So that it can be queried and removed properly.

Bug: n/a
Test: LayoutNodeMappingTest
Change-Id: I9eee477c0e4416678c464f4bc09316cda891aeea