Add base classes for Nav3
Adding the building blocks for the nav3 library. The Record class
provides a connection between a key and some composable content. The
NavContentWrapper allows passing of information to either the back stack
or a record. The NavWrapperManager handles the interactions between the
list of NavContentWrappers and then currently active content.

RelNote: N/A
Test: Added RecordTest
Bug: 374890821
Change-Id: Ifa0d882e1930c8c438667652ed0e496ea8fc1df1