Change ExifInterface's handling of XMP in HEIC files
The previously documented behaviour of preferring the XMP from inside
the Exif data, rather than from a standalone segment in the file,
doesn't seem aligned with the XMP spec. The previous behaviour has only
been released in 1.4.0-alpha01, so we can change it now before
1.4.0-beta01.

This change also removes some duplicated javadoc that should have been
removed as part of http://r.android.com/3140676.

I tried to craft a HEIC file with Exif in both parts, in order to test
this change, but I couldn't convince exiftool to do it, and since
ExifInterface doesn't support writing HEIC I couldn't hack it to do it
either (like I did for the test JPEG files with XMP in both places).

Test: ExifInterfaceTest

Change-Id: If90ede190cff7e8f2a5b3232b186c7e757582b2e