Allow subcomposed items to do a make-up lookahead placement
Setup a separate code path to allow subcomposed items to skip
lookahead placement and go straight to approach placement. When
that happens, a make-up lookahead placement will take place
first before the approach placement for the item. This is to
obey the contract that for any layout nodes, lookahead will
happen before approach.

This make-up placement is achieved by creating a separate
flag for when the placement of a child is detached from
its parent's lookahead pass. In that case, the approach
placement becomes the trigger for the particular child. The
`placed` tracking has been updated to reflect the complexity
of placing a lookahead delegate during approach.

Note: This support for skipping lookahead placement is scoped
to only direct children of SubcomposeLayout, as we:
1) don't expect non-subcomposeLayout to skip lookahead placement,
and 2) want to avoid masking potential issues

TODO: Check detachedFromParentLookaheadFlag when invalidating lookahead
placement

Test: New tests added

Change-Id: I249d77eaa98bd132289b3d26a854eefd227dcc5e