Fix multi-instance invalidation for Flow
On aosp/3264071, the InvalidationTracker Flow APIs were refactored and their implementation are no longer backed by an InvalidationTracker.Observer, an Android only API as the newer APIs are KMP compatible. This broke multi-instance invalidation since MultiInstanceInvalidationClient relied on notifyObserversByTableNames() to notify observers.

This change fixes the issue by merging an additional invalidation Flow created from the MultiInstanceInvalidationClient when multi-instance invalidation is enabled such that DAO Flows or those created directly with the InvalidationTracker emit due to changes coming through the MultiInstanceInvalidationService and specifically another instance.

The MultiInstanceInvalidationClient is no longer set to null and re-instantiated during auto-close thanks to the changes in aosp/3159516 where binding occurs in start(). This meant we can remove MultiInstanceClientInitState and initialize the client early if the feature is enabled. This was important since otherwise the client is only initialize when the database is open and because creations don't require opening a connection it means we need to have a client early in case we need to merge the flows.

Bug: 377972487
Test: MultiInstanceInvalidationTest
Change-Id: Ie5e4aae8ae407b5e0e8f0f8d76a7ec5f5f9c4297