ExifInterface: Update Exif offset in saveAttributes
This ensures that methods like ExifInterface.getThumbnail (which rely
on mOffsetToExifData) continue to work after saveAttributes.

These methods (unlike getAttributeRange) are not documented to fail
after saveAttributes, and the current behaviour is undefined if
saveAttributes resulted in the Exif data moving in the underlying file.
It can be an EOFException resulting in returning null, or just returning
garbage bytes that can't be decoded to a Bitmap.

This change re-purposes the (previously unused) return value of the
private writeExifSegment method to indicate the offset into the
parameter output stream where the Exif data starts.

We currently have tests that exercise getThumbnailBytes after
saveAttributes, but none of our tests result in the Exif segment moving
because all our test images have the APP1 segment at the start, which is
also where ExifInterface writes it. This change includes a new test with
a JPEG image that has the APP1 segment later, and this gets moved to the
start of the file during saveAttributes, which changes the offsets and
breaks invocations of getThumbnailBytes without this change.

Test: ExifInterfaceTest
Change-Id: Id482f1ad193966e1301a9514ab673a7b72539e26