Fix performance regression caused by rememberPaneExpansionState
The regression is caused by the introduction of LaunchEffect in the
remember function, which is required to keep the state correctly between
scaffold value changes.

However, when paneExpansionDragHandle is not provided, there's no way
for devs to mutate the default paneExpansionState. In this case, we can
avoid a lot of management overhead by just providing a stub
implementation.

Relnote: Makes pane expansion state null by default
Test: benchmarks and existing instrumentation tests
Bug: 376394520
Change-Id: Ia65f8919ba9aaaa4c611ecb40514ee41d1e8d56f