[Bottomsheet] Update modifier calls to add user Modifier after Material modifiers.
This is specifically needed for a11y focus, which needs to be implemented after the draggable modifier to respond to size and position changes.

Bug: 376814403
Test: Manual
RelNote: modifier pararemeter is now applied after internal modifiers such as dragging behavior and semantics, instead of before. This affects the ordering in which modifiers are applied.
Change-Id: I8d83fd447173da5cd76aa6e03ca2d0b4921c7a59