Update NonLambdaOffsetModifierDetector to catch qualified references in delegate declarations
This lint check inspects delegated properties to see if they are stateful, but it was only able to find the type for a delegate created with a `UCallExpression`, like `by remember {...}`. This didn't work for other types of expressions, like `by androidx.compose.runtime.remember {...}`, which is a `UQualifiedReferenceExpression`.

This came up because a new version of android lint improves the FULLY_QUALIFIED test mode to fully qualify `remember` which it previously was not fully qualifying, which would cause some of these tests to fail.

Bug: 378110980
Test: Added qualified delegated property case to NonLambdaOffsetModifierDetectorTest
Change-Id: I6801d875990e2aa45a621308cd9f5f2b020dca07