Move managedDevices.devices to allDevices
Gradle Managed Devices is moving to deprecate managedDevices.devices.
It was replaced by managedDevices.allDevices (for any GMD device type) and localDevices (specifically for ManagedVirtualDevices) as an effort to simplify the managed device DSL back in 2023. The implementation is the same for both devices and allDevices, (their interfaces point to the same domain object container) so no other changes are necessary.

This change will ensure that androidx is not using the deprecated api.

Bug: 279042301
Change-Id: I951b8382ce2770f988b637dc0e1aab6d4265874c
Tests: existing