Fix NavDestination mismatch when adding NavBackStackEntry
When navigating to a nested NavDestination, we trace the target destination's hierarchy and add its parents to the backstack. Since a destination can have different parents (i.e. same destination exists within different graphs), we check its parent for referential equality and add it only if that specific parent hasn't been added yet.

Part of this process includes searching for a specific parent instance. However, parents can also be duplicated - a parent destination can also exist in several different graphs. The current bug is due to findDestination returning the wrong parent duplicate.

So findDestination is modified to take an optional matchingDestination parameter. When matchingDestination is present, findDestination will attempt to return a destination that is equal (==) and has the same parent (same nesting) as the matchingDestination.

Bug: 361560785
Test: ./gradlew navigation:navigation-runtime:cC
Relnote: "Fixed NavDestination NullPointerException when updating a Graph's startDestination"
Change-Id: I99421bf1e4aef9802873d9d78fbc6fafc15a21ba