Made arithmatic on `SnapshotId` internal.
Added a `toLong()` to the `SnapshotId` to allow it to losslessly
be converted to a type that supports arithmatic.

Fixes: 378544379
Test: existing tests, ./gradlew :compose:r:r:test
Relnote: """Made the arithmatic and special constants for `SnapshotId`
internal instead of public. Arithmetic can be performed on a
`SnasphotId`, if necessary, by converting it to a `Int` or `Long`
by using `toInt()` or `toLong()` respectively."""

Change-Id: Ic3a57eed3968234376877c743a23dc13a127fbe0