Use pausable composition in prefetch to split composition and apply
As a first step we only partially use the new functionality without really ever pausing the composition, but taking an advantage from the fact that we can do composition and apply as separately.

Test: existing benchmarks
Relnote: PausableComposition is now used during the lazy layout prefetch. There is a new feature flag allowing to disable it if it causes any issues ComposeFoundationFlags.isPausableCompositionInPrefetchEnabled. Please file a bug if it does.
Change-Id: I9d71ce536983f985cfb27837cfdaf8b6a6014cdd