Fix exception not thrown when profile not bundled
In `DeviceProfileWriter#openStreamFromAssets` `FileNotFoundException` is
caught in order to check if the message contains the word `compressed`.
If not, the exception is not rethrown and as a result
`FileNotFoundException` is never thrown. This method is used both when
reading metadata and the profile so the exception is not called in both
these cases. The fix simply remove the `openStreamFromAssets` and
evaluates whether the file not found should be thrown for the profile.

Fix:379429942
Test: Existing
Change-Id: I98b8ccbb1497a77a8be89154708efbdba3ea1659