Fix a bug where CameraDeviceWrapper isn't given to Camera2DeviceCloser
In this section, AndroidCameraState has already emitted CameraStateOpen,
which means capture sessions could technically happen before we close
the camera device. Not providing CameraDeviceWrapper may lead to capture
session states not getting finalized. This is a highly unplausible
scenario, but fixing this as part of a code trace and cleanup to be done
later.

Bug: 379347826
Test: OpenCloseCameraStressTest on Pixel
Change-Id: I506077ac5b791d6da2808c43b095e4b9ecf593c0