Make UiLib Test App compatible with EdgeToEdge Mode
Adds changes to the UiLib TestApp's theme and gradle
build files to make it compatible with edge to edge
mode introduced in Android 15+ devices. The test app
will expand to the boundaries of the entire display
now and the publisher can set margins or paddings by
listening to the system bars insets to make the app
UI more user friendly. Also removes an empty
SandboxDeathFragment file.

Bug: 379877120
Test: UiLib Test App
Change-Id: I62ceaa456eb1b2b52e28248f0cfa87f5b351090e