Create empty capture session with CameraDeviceWrapper
This CL migrates the creation of empty capture session from using
CameraDevice directly to using CameraDeviceWrapper. This is to ensure
capture session switching on any camera is done locally in one place.

Note if we don't have a CameraDeviceWrapper, that means the camera
wasn't passed down and therefore wasn't used. In that scenario, an empty
capture session wouldn't have been needed in the first place.

Bug: 379347826
Test: OpenCloseCameraStressTest on legacy devices
Change-Id: Ied120f2dd9d044e451b154c3425cada3067384b5