Remove the dependency on XR extensions AAR
The xr.compose.material3.integration-tests.testapp was
depending on the androidx.xr.extensions-release.aar which
includes the androidx.xr.extensions API. This API has
been moved to the SceneCore library which this test app
also depends on. That caused duplicate class build errors.

This change set removes the dependency on the AAR.

Bug: 370101400
Change-Id: I299bd3608e95cdfc2176a06060ed8367d1f69745
Test: Built and deployed the testapp on device.