Additional fix for drag gesture pick up.
In some cases the we may receive an unconsumed event in the middle of consumed events. This means we were giving up on the gesture pick up if we receive one of these one off events. We can continue trying to pick it up until all fingers are up or if the gesture was consumed by someone down the line.

Test: Additional tests.
Bug: 303224944
Change-Id: I7d0168d98a9001f8523b001c1763403219e33102