Do not process state closures if the state is closed.
When the camera state is closed due to an error, it is possible we
update the state of the controller even after it is closed. This makes
it possible for camera controller to attempt a restart if the state was
erroneously updated to ERROR.

This CL makes early returns such that we do not process state closures
if the state is closed, and add a unit test for such cases.

Bug: 380282226
Test: CameraPipe test app.
Change-Id: I96dafdbb6291dfdac415a86b55f257710cb2debf