Move media3 effect artifact to new folder.
We need to create a new group name for the artifact, because it cannot move to stable with the rest of the CameraX artifacts. As per suggestion from the AndroidX team, we will also move the code location based on the output of ./development/project-creator/create_project.sh.

Relnote: change media3effect library group name and create new code path.

Bug: 380310983
Test: manual test and ./gradlew bOS
Change-Id: I65ca4f0e2fa29278fb96d892aacc489897b02417