Update dynamicColorScheme API
Removed defaultColorScheme parameter, which always created a ColorScheme() object and return nullable ColorScheme instead, so that the caller must handle the fallback explicitly.

Bug: 380278095
Test: Updated integration demo and tested on device
Relnote: "We have updated dynamicColorScheme by removing the optional defaultColorScheme parameter and now returning nullable ColorScheme. This means that the caller must explicitly handle the fallback case when a dynamic color scheme is not provided."
Change-Id: I6d62ea8d01175b2a95772b771a469bed96071231