Add further capabilities to WV Async Startup API
This CL adds two new capabilities to the Async Startup
API when the underlying WebView supports it:
- Diagnostic information about WebView startup is provided
to the app.
- Provide an option for apps to only trigger provider init.

Appropriate unit tests are written to test the new
code. This completes the current iteration of the API as
documented in the below design doc.

Design doc: go/wv-experimental-async-startup-api

Test: ./gradlew webkit:integration-tests:testapp:connectedAndroidTest
Bug: 374304951
Relnote: "- Get diagnostic information from WebView async startup API.
- Provide an option for apps to only trigger init that doesn't
block the UI thread."

Change-Id: I9bf2b3b53b835366cf68c938a8bcd1612bd7f3fd