Implement tiled rendering at high zoom levels
Separates the concern of fetching bitmaps from Page to a new BitmapFetcher

Page remains responsible for drawing bitmaps as well as for relaying state changes to BitmapFetcher

Notably there is some room for memory optimization, tracked in b/382558660. Functionally and latency-wise this seems to be at parity with the previous implementation.

All "tuning" values (tile size, threshold for tiled rendering, etc) are adapted from the previous implementation.

Bug: 376135536
Test: ./gradlew :pdf:pdf-viewer:(test|cAT), manual
Change-Id: I75becbdd49736a8b48e7c41fc74979e02735462b