Add sample project for Tiles
It's way quicker turnaround if we have samples app to run tile (as a tile)
in AndroidX instead of relying on screenshot tests and waiting on
results to verify changes.

Later on, we can add more tiles, or functionality to this tile to show
multiple layouts.

But it can also be used as a local playground to test changes.
- App: https://screenshot.googleplex.com/6yvqpG3ET584ZgT.png
- Tile: https://screenshot.googleplex.com/rTUtdVu9oyrQmAZ.png

Fixes: 381442729
Test: Run locally, AX Playground Tile

Change-Id: Ibe8d9cf9430d1ea3e97364325e47a158eaa19b62